split host into host and interfaces
