+++ openpype/host/__init__.py
@@ -1,5 +1,8 @@
 from .host import (
     HostBase,
+)
+
+from .interfaces import (
     IWorkfileHost,
     ILoadHost,
     INewPublisher,
@@ -7,6 +10,7 @@ from .host import (
 
 __all__ = (
     "HostBase",
+
     "IWorkfileHost",
     "ILoadHost",
     "INewPublisher",
